Part 5.6
6. Add a new field boolean available in Book (Default value should be true).
Make sure you set available to false when a BookLoan is added to AppUser.
Also make sure you can NOT lend a book that has available set to false.

Added Field as point 6. and of course tests that go with it.
Added field and managed the add and remove methods.